chore: remove meteor.startup from permalink-star
