+++ apps/meteor/client/components/message/toolbar/MessageToolbar.tsx
@@ -11,6 +11,7 @@ import React, { memo, useMemo, useRef } from 'react';
 import MessageActionMenu from './MessageActionMenu';
 import MessageToolbarStarsActionMenu from './MessageToolbarStarsActionMenu';
 import { useNewDiscussionMessageAction } from './useNewDiscussionMessageAction';
+import { usePermalinkStar } from './usePermalinkStar';
 import { useStarMessageAction } from './useStarMessageAction';
 import { useUnstarMessageAction } from './useUnstarMessageAction';
 import { useWebDAVMessageAction } from './useWebDAVMessageAction';
@@ -92,6 +93,7 @@ const MessageToolbar = ({
 	useNewDiscussionMessageAction();
 	useStarMessageAction(message, { room, user });
 	useUnstarMessageAction(message, { room, user });
+	usePermalinkStar();
 
 	const actionsQueryResult = useQuery({
 		queryKey: roomsQueryKeys.messageActionsWithParameters(room._id, message),
