+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -3,6 +3,5 @@ import './jumpToPinMessage';
 import './jumpToSearchMessage';
 import './jumpToStarMessage';
 import './permalinkPinned';
-import './permalinkStar';
 import './pinMessage';
 import './unpinMessage';
